GetMention
A standardized way to get mentions out of a message. An example of its usage might look like this:
```php
$mentions_arr 	= $message->mentions->users;
$GetMentionResult = GetMention([$guild, strtolower($message_content), $filter, $option]);
if (!(is_array($GetMentionResult))) return $message->reply("Invalid input! Please enter a valid ID or @mention the user");
if ($GetMentionResult[1] == NULL) return $message->reply("User not found in the guild!");
$mention_user = GetMentionResult[0];
$mention_member = GetMentionResult[1];
$mentions_arr = $mentions_arr ?? GetMentionResult[2];
```